moved wellknownendpoints and fixed some includes
