+++ fdbclient/CoordinationInterface.h
@@ -27,7 +27,7 @@
 #include "fdbrpc/Locality.h"
 #include "fdbclient/CommitProxyInterface.h"
 #include "fdbclient/ClusterInterface.h"
-#include "fdbclient/WellKnownEndpoints.h"
+#include "fdbrpc/WellKnownEndpoints.h"
 #include "flow/Hostname.h"
 
 const int MAX_CLUSTER_FILE_BYTES = 60000;
