+++ fdbclient/DatabaseBackupAgent.actor.cpp
@@ -26,7 +26,7 @@
 #include "fdbclient/NativeAPI.actor.h"
 #include <ctime>
 #include <climits>
-#include "fdbrpc/IAsyncFile.h"
+#include "flow/IAsyncFile.h"
 #include "flow/genericactors.actor.h"
 #include "flow/Hash3.h"
 #include <numeric>
