+++ fdbrpc/AsyncFileEncrypted.h
@@ -20,7 +20,7 @@
 
 #pragma once
 
-#include "fdbrpc/IAsyncFile.h"
+#include "flow/IAsyncFile.h"
 #include "flow/FastRef.h"
 #include "flow/flow.h"
 #include "flow/IRandom.h"
