+++ fdbrpc/HTTP.h
@@ -25,8 +25,7 @@
 
 #include "flow/flow.h"
 #include "flow/Net2Packet.h"
-#include "fdbrpc/IRateControl.h"
-#include "fdbclient/Knobs.h"
+#include "flow/IRateControl.h"
 
 namespace HTTP {
 struct is_iless {
