+++ fdbrpc/Net2FileSystem.h
@@ -23,7 +23,7 @@
 #include <string>
 #pragma once
 
-#include "fdbrpc/IAsyncFile.h"
+#include "flow/IAsyncFile.h"
 
 class Net2FileSystem final : public IAsyncFileSystem {
 public:
