+++ fdbserver/RestoreUtil.h
@@ -32,7 +32,7 @@
 #include "flow/flow.h"
 #include "fdbrpc/TimedRequest.h"
 #include "fdbrpc/fdbrpc.h"
-#include "fdbrpc/IAsyncFile.h"
+#include "flow/IAsyncFile.h"
 #include "fdbrpc/Stats.h"
 #include <cstdint>
 #include <cstdarg>
