+++ fdbserver/RestoreWorker.actor.h
@@ -28,7 +28,7 @@
 #include "fdbclient/Tuple.h"
 #include "flow/flow.h"
 #include "fdbrpc/fdbrpc.h"
-#include "fdbrpc/IAsyncFile.h"
+#include "flow/IAsyncFile.h"
 #include "fdbrpc/Stats.h"
 #include <cstdint>
 #include <cstdarg>
