+++ fdbserver/VFSAsync.h
@@ -21,7 +21,7 @@
 #include "sqlite/sqlite3.h"
 #include <string>
 #include <map>
-#include "fdbrpc/IAsyncFile.h"
+#include "flow/IAsyncFile.h"
 #include "fdbrpc/simulator.h"
 
 /*
