+++ fdbserver/VersionedBTree.actor.cpp
@@ -38,7 +38,7 @@
 #include "flow/genericactors.actor.h"
 #include "flow/UnitTest.h"
 #include "fdbserver/IPager.h"
-#include "fdbrpc/IAsyncFile.h"
+#include "flow/IAsyncFile.h"
 #include "flow/ActorCollection.h"
 #include <map>
 #include <string>
