+++ fdbserver/workloads/AsyncFile.actor.h
@@ -29,7 +29,7 @@
 #define WORKLOADS_ASYNCFILE_ACTOR_H
 
 #include "fdbserver/workloads/workloads.actor.h"
-#include "fdbrpc/IAsyncFile.h"
+#include "flow/IAsyncFile.h"
 #include "flow/actorcompiler.h" // This must be the last #include.
 
 class RandomByteGenerator {
