+++ fdbserver/workloads/AsyncFileRead.actor.cpp
@@ -23,7 +23,7 @@
 #include "fdbserver/workloads/workloads.actor.h"
 #include "flow/ActorCollection.h"
 #include "flow/SystemMonitor.h"
-#include "fdbrpc/IAsyncFile.h"
+#include "flow/IAsyncFile.h"
 #include "fdbserver/workloads/AsyncFile.actor.h"
 #include "flow/DeterministicRandom.h"
 #include "flow/actorcompiler.h" // This must be the last #include.
